Relocated AP signature functions, reduced magic auth functions
